feat: add buzzer module and integrate with main project
