+++ CMakeLists.txt
@@ -1,17 +1,3 @@
-# == DO NOT EDIT THE FOLLOWING LINES for the Raspberry Pi Pico VS Code Extension to work ==
-if(WIN32)
-    set(USERHOME $ENV{USERPROFILE})
-else()
-    set(USERHOME $ENV{HOME})
-endif()
-set(sdkVersion 2.2.0)
-set(toolchainVersion 14_2_Rel1)
-set(picotoolVersion 2.2.0)
-set(picoVscode ${USERHOME}/.pico-sdk/cmake/pico-vscode.cmake)
-if (EXISTS ${picoVscode})
-    include(${picoVscode})
-endif()
-# ====================================================================================
 cmake_minimum_required(VERSION 3.13)
 set(PICO_BOARD pico_w CACHE STRING "Board type")
 set(CMAKE_C_STANDARD 11)
@@ -31,6 +17,7 @@ add_executable(${PROJECT_NAME}
         lib/rgb/rgb.c
         lib/push_button/button.c
         lib/matriz/matriz.c
+        lib/buzzer/buzzer.c
         )
 
 
@@ -55,6 +42,7 @@ target_include_directories(${PROJECT_NAME} PRIVATE ${CMAKE_CURRENT_LIST_DIR}
             lib/rgb
             lib/push_button
             lib/matriz
+            lib/buzzer
         )
 
 pico_enable_stdio_usb(${PROJECT_NAME} 1)
